fix protocol index
